Exit with test status code

+++ run-integration-tests.sh
@@ -30,30 +30,30 @@ cargo build --test webdriver
 INTEGRATION_TEST_BINARY=$(cargo build --test webdriver --message-format json | jq --raw-output 'select(.reason == "compiler-artifact" and .target.name == "webdriver") | .executable')
 echo "Compiled integration test binary: $INTEGRATION_TEST_BINARY"
 
-: '
-podman network create --ignore pga
+if [ "${1-}" == "keycloak" ]; then
+    podman network create --ignore pga
 
-podman build -t keycloak --file keycloak/keycloak/Dockerfile .
-podman build -t perfect-group-allocation --file base/perfect-group-allocation/Dockerfile .
-podman build -t test --file base/test/Dockerfile .
+    podman build -t keycloak --file keycloak/keycloak/Dockerfile .
+    podman build -t perfect-group-allocation --file base/perfect-group-allocation/Dockerfile .
+    podman build -t test --file base/test/Dockerfile .
 
-(cd keycloak && CAROOT=$CAROOT mkcert keycloak)
-(cd keycloak && kustomize edit set nameprefix tmp-)
-(cd keycloak && kustomize edit add patch --patch '{"apiVersion": "v1","kind": "Pod","metadata":{"name":"keycloak"},"spec":{"volumes":[{"name":"root-ca","hostPath":{"path":"'"$CAROOT"'/rootCA.pem"}}]}}')
-(cd keycloak && kustomize build --output kubernetes.yaml)
-(cd keycloak && podman kube down --force kubernetes.yaml || exit 0) # WARNING: this also removes volumes
-(cd keycloak && podman kube play --network pga kubernetes.yaml)
-echo waiting for keycloak
-podman wait --condition healthy tmp-keycloak-keycloak
-echo keycloak started
-podman exec tmp-keycloak-keycloak keytool -noprompt -import -file /run/rootCA.pem -alias rootCA -storepass password -keystore /tmp/.keycloak-truststore.jks
-podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh config truststore --trustpass password /tmp/.keycloak-truststore.jks
-podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh config credentials --server https://keycloak --realm master --user admin --password admin
-podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh create realms -s realm=pga -s enabled=true
-podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh create users -r pga -s username=test -s email=test@example.com -s enabled=true
-podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh set-password -r pga --username test --new-password test
-podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh create clients -r pga -s clientId=pga -s secret=$(cat base/client-secret) -s 'redirectUris=["https://perfect-group-allocation/openidconnect-redirect"]'
-'
+    (cd keycloak && CAROOT=$CAROOT mkcert keycloak)
+    (cd keycloak && kustomize edit set nameprefix tmp-)
+    (cd keycloak && kustomize edit add patch --patch '{"apiVersion": "v1","kind": "Pod","metadata":{"name":"keycloak"},"spec":{"volumes":[{"name":"root-ca","hostPath":{"path":"'"$CAROOT"'/rootCA.pem"}}]}}')
+    (cd keycloak && kustomize build --output kubernetes.yaml)
+    (cd keycloak && podman kube down --force kubernetes.yaml || exit 0) # WARNING: this also removes volumes
+    (cd keycloak && podman kube play --network pga kubernetes.yaml)
+    echo waiting for keycloak
+    podman wait --condition healthy tmp-keycloak-keycloak
+    echo keycloak started
+    podman exec tmp-keycloak-keycloak keytool -noprompt -import -file /run/rootCA.pem -alias rootCA -storepass password -keystore /tmp/.keycloak-truststore.jks
+    podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh config truststore --trustpass password /tmp/.keycloak-truststore.jks
+    podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh config credentials --server https://keycloak --realm master --user admin --password admin
+    podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh create realms -s realm=pga -s enabled=true
+    podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh create users -r pga -s username=test -s email=test@example.com -s enabled=true
+    podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh set-password -r pga --username test --new-password test
+    podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh create clients -r pga -s clientId=pga -s secret=$(cat base/client-secret) -s 'redirectUris=["https://perfect-group-allocation/openidconnect-redirect"]'
+fi
 
 (cd base && CAROOT=$CAROOT mkcert perfect-group-allocation)
 (cd base && kustomize edit set nameprefix tmp-)
@@ -63,4 +63,5 @@ podman exec tmp-keycloak-keycloak /opt/keycloak/bin/kcadm.sh create clients -r p
 (cd base && kustomize build --output kubernetes.yaml)
 (cd base && podman kube down --force kubernetes.yaml || exit 0) # WARNING: this also removes volumes
 (cd base && podman kube play --network pga kubernetes.yaml)
-podman logs --color --names --follow tmp-test-test # tmp-perfect-group-allocation-perfect-group-allocation # tmp-keycloak-keycloak tmp-postgres-postgres 
+podman logs --color --names --follow tmp-test-test tmp-perfect-group-allocation-perfect-group-allocation & # tmp-keycloak-keycloak tmp-postgres-postgres 
+exit $(podman wait tmp-test-test)
